MAINT: rename solver -> groundstate
