+++ smol/capp/generate/groundstate/__init__.py
@@ -1,0 +1,1 @@
+"""Ground state groundstate functionalities."""
