+++ smol/capp/generate/groundstate/upper_bound/__init__.py
