--- smol/capp/generate/solver/__init__.py
@@ -1,1 +1,0 @@
-"""Ground state solver functionalities."""
