+++ tests/test_solver/test_constraints.py
@@ -4,8 +4,8 @@ import itertools
 import pytest
 from pymatgen.core import Species
 
-from smol.capp.generate.solver import get_upper_bound_variables_from_sublattices
-from smol.capp.generate.solver.upper_bound.constraints import (
+from smol.capp.generate.groundstate import get_upper_bound_variables_from_sublattices
+from smol.capp.generate.groundstate.upper_bound.constraints import (
     get_upper_bound_composition_space_constraints,
     get_upper_bound_fixed_composition_constraints,
     get_upper_bound_normalization_constraints,
