+++ tests/test_solver/test_objectives.py
@@ -2,17 +2,17 @@
 import numpy as np
 import numpy.testing as npt
 
-from smol.capp.generate.solver import (
+from smol.capp.generate.groundstate import (
     get_occupancy_from_variables,
     get_upper_bound_variables_from_sublattices,
 )
-from smol.capp.generate.solver.upper_bound.objectives import (
+from smol.capp.generate.groundstate.upper_bound.objectives import (
     get_upper_bound_terms_from_chemical_potentials,
     get_upper_bound_terms_from_decomposition_processor,
     get_upper_bound_terms_from_ewald_processor,
     get_upper_bound_terms_from_expansion_processor,
 )
-from smol.capp.generate.solver.upper_bound.utils.terms import (
+from smol.capp.generate.groundstate.upper_bound.utils.terms import (
     get_auxiliary_variable_values,
     get_expression_and_auxiliary_from_terms,
 )
